# apps/api/src/modules/tags/dto/tag.dto.ts
import { ApiProperty } from "@nestjs/swagger"

export class TagDto {
  @ApiProperty()
  id!: string

  @ApiProperty()
  name!: string

  @ApiProperty()
  color!: string

  @ApiProperty()
  icon!: string
}